+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Change Log
 
-All notable changes to the "vscode-editor-env" extension will be documented in this file.
+All notable changes to the pathfinder extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
-## [Unreleased]
+## [1.0.0]
 
 - Initial release
+
+### Added
+
+- Writes workspace folder to `/tmp/folder_path.txt`, defaults to `~`
+- Editor path now supports non-text editors
+
+### Changed
+
+- `/tmp/editor_path.txt` now defaults to the workspace folder
+
+## [0.0.1]
+
+- Includes support for `/tmp/editor_path.txt`